Split up DSL into resource/endpoint/callback concerns. Ref #28
No attempt was made to clean up the individual DSLs, but this should be a decent starting point to better define the boundaries.
